+++ packages/sdk/rslib.config.ts
@@ -15,17 +15,13 @@ export default defineConfig({
     {
       format: 'esm',
       syntax: ['node 22'],
-      output: {
-        minify: true,
-      },
+      output: { minify: true },
       dts: true,
     },
     {
       format: 'cjs',
       syntax: ['node 22'],
-      output: {
-        minify: true,
-      },
+      output: { minify: true },
     },
   ],
 });