+++ packages/sdk/src/accounts.ts
@@ -4,7 +4,7 @@ export type {
 } from '@bloque/sdk-accounts';
 
 import { AccountsClient } from '@bloque/sdk-accounts';
-import { getHttpClient } from './http';
+import { getHttpClient } from './config';
 import { lazyClient } from './lazy';
 
 export const Accounts = lazyClient<AccountsClient>(