+++ packages/sdk/src/compliance.ts
@@ -5,7 +5,7 @@ export type {
 } from '@bloque/sdk-compliance';
 
 import { ComplianceClient } from '@bloque/sdk-compliance';
-import { getHttpClient } from './http';
+import { getHttpClient } from './config';
 import { lazyClient } from './lazy';
 
 export const Compliance = lazyClient<ComplianceClient>(