+++ packages/sdk/src/identity.ts
@@ -1,7 +1,7 @@
 export type { Alias } from '@bloque/sdk-identity';
 
 import { IdentityClient } from '@bloque/sdk-identity';
-import { getHttpClient } from './http';
+import { getHttpClient } from './config';
 import { lazyClient } from './lazy';
 
 export const Identity = lazyClient<IdentityClient>(