+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @bloque/sdk-core
 
+## 0.0.20
+
+### Patch Changes
+
+- dummy
+
 ## 0.0.19
 
 ### Patch Changes